+++ projects/com.nvidia.grcuda.test/src/com/nvidia/grcuda/test/cudalibraries/CUMLTest.java
@@ -50,7 +50,7 @@ public class CUMLTest {
     public static Collection<Object[]> data() {
 
         return GrCUDATestUtil.crossProduct(Arrays.asList(new Object[][]{
-                        {ExecutionPolicyEnum.SYNC.getName(), ExecutionPolicyEnum.ASYNC.getName()},
+                        {ExecutionPolicyEnum.SYNC.toString(), ExecutionPolicyEnum.ASYNC.toString()},
                         {true, false},
                         {'S', 'D'}
         }));