+++ projects/com.nvidia.grcuda/src/com/nvidia/grcuda/functions/GetOptionsFunction.java
@@ -28,20 +28,26 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.nvidia.grcuda.runtime.stream;
 
-public enum RetrieveParentStreamPolicyEnum {
-    SAME_AS_PARENT("same-as-parent"),
-    DISJOINT("disjoint");
+package com.nvidia.grcuda.functions;
 
-    private final String name;
+import com.nvidia.grcuda.GrCUDAOptionMap;
+import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
+import com.oracle.truffle.api.interop.ArityException;
+import com.oracle.truffle.api.interop.UnsupportedTypeException;
 
-    RetrieveParentStreamPolicyEnum(String name) {
-        this.name = name;
+public class GetOptionsFunction extends Function {
+    private final GrCUDAOptionMap optionMap;
+
+    public GetOptionsFunction(GrCUDAOptionMap optionMap) {
+        super("getoptions");
+        this.optionMap = optionMap;
     }
 
     @Override
-    public String toString() {
-        return name;
+    @TruffleBoundary
+    public Object call(Object[] arguments) throws UnsupportedTypeException, ArityException {
+        checkArgumentLength(arguments, 0);
+        return optionMap;
     }
 }