+++ projects/com.nvidia.grcuda/src/com/nvidia/grcuda/runtime/computation/dependency/DependencyPolicyEnum.java
@@ -40,7 +40,8 @@ public enum DependencyPolicyEnum {
         this.name = name;
     }
 
-    public String getName() {
+    @Override
+    public String toString() {
         return name;
     }
 }