+++ projects/com.nvidia.grcuda/src/com/nvidia/grcuda/runtime/stream/RetrieveNewStreamPolicyEnum.java
@@ -40,7 +40,8 @@ public enum RetrieveNewStreamPolicyEnum {
         this.name = name;
     }
 
-    public String getName() {
+    @Override
+    public String toString() {
         return name;
     }
 }